requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
argparse==1.4.0
    # via -r requirements.in
certifi==2021.5.30
    # via
    #   httpx
    #   requests
chardet==3.0.4
    # via httpx
charset-normalizer==2.0.6
    # via requests
click==8.0.1
    # via
    #   nltk
    #   sacremoses
contextvars==2.4
    # via sniffio
dataclasses
    # via
    #   torch
    #   transformers
filelock==3.3.0
    # via
    #   huggingface-hub
    #   transformers
fuzzywuzzy==0.18.0
    # via parrot
googletrans==3.1.0a0
    # via -r requirements.in
googletranslate==2.1.2
    # via -r requirements.in
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2021.10.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via googletrans
huggingface-hub==0.0.17
    # via
    #   sentence-transformers
    #   transformers
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   requests
immutables==0.16
    # via contextvars
importlib-metadata==4.8.1
    # via
    #   click
    #   huggingface-hub
    #   transformers
inflect==5.3.0
    # via -r requirements.in
joblib==1.0.1
    # via
    #   nltk
    #   sacremoses
    #   scikit-learn
leven==1.0.4
    # via -r requirements.in
nltk==3.6.3
    # via
    #   -r requirements.in
    #   sentence-transformers
nose==1.3.7
    # via leven
numpy==1.19.5
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   symspellpy
    #   torchvision
    #   transformers
packaging==21.0
    # via
    #   huggingface-hub
    #   transformers
pandas==1.1.5
    # via -r requirements.in
parrot @ git+https://github.com/PrithivirajDamodaran/Parrot.git
    # via -r requirements.in
pillow==8.3.2
    # via torchvision
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.2
    # via pandas
python-levenshtein==0.12.2
    # via parrot
pytz==2021.3
    # via pandas
pyyaml==5.4.1
    # via transformers
regex==2021.9.30
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.26.0
    # via
    #   googletranslate
    #   huggingface-hub
    #   transformers
rfc3986==1.5.0
    # via httpx
sacremoses==0.0.46
    # via transformers
scikit-learn==0.24.2
    # via sentence-transformers
scipy==1.5.4
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.1.0
    # via parrot
sentencepiece==0.1.96
    # via
    #   -r requirements.in
    #   parrot
    #   sentence-transformers
six==1.16.0
    # via
    #   leven
    #   python-dateutil
    #   sacremoses
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
symspellpy==6.7.0
    # via -r requirements.in
threadpoolctl==3.0.0
    # via scikit-learn
tokenizers==0.10.3
    # via
    #   sentence-transformers
    #   transformers
torch==1.9.1
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   torchvision
torchvision==0.10.1
    # via sentence-transformers
tqdm==4.62.3
    # via
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   sentence-transformers
    #   transformers
transformers==4.11.2
    # via
    #   -r requirements.in
    #   parrot
    #   sentence-transformers
typing-extensions==3.10.0.2
    # via
    #   huggingface-hub
    #   immutables
    #   importlib-metadata
    #   torch
urllib3==1.26.7
    # via requests
zipp==3.6.0
    # via importlib-metadata
matplotlib
datasketch
jupyter
requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
